Re-queue jobs if the child is terminated
If the child process was killed with a SIGTERM or SIGINT, there is
likely nothing wrong with the job. Instead of failing these jobs, put
them back on the head of their queue.  This lets us more safely kill
workers without having to worry about orphaned jobs, or unintended
failures.

Raise KilledError to signal the the worker was killed abruptly, and
requeue the job. Worker.requeue uses an lpush to put it back at the head
of the queue for immediate reservation, and cleans up without marking
the job as done, or raising further exceptions.